[Feat] : 서버 오브젝트 위치 값을 정수형(UINT16)에서 실수형(float)으로 변경

+++ 3DFPSServer/Content.cpp
@@ -55,7 +55,7 @@ void LoginAcceptObject(CObject* pObj)
     //=====================================================================================================================================
 
     // 룸에 있던 activePlayer들의 정보를 새로 룸에 접속한 플레이어에게 전송
-    UINT16 posX, posY, posZ;
+    float posX, posY, posZ;
     UINT32 rotationAxisX, rotationAxisY;
     KDAInfo kdaInfo;
     for (const auto& activePlayer : room->m_activePlayers)
